edit the types and remove duplicates
