+++ src/pages/Monitoring/monitoringServices.ts
@@ -1,5 +1,6 @@
 import { httpClient, getHeaders } from "../../services/api/httpClient";
 import { API } from "../../constants/ConstantKeys.constants";
+import type { PaginationMeta } from "../../types/types";
 
 export interface GetSessionItemInterface {
   id: number;
@@ -15,39 +16,6 @@ export interface GetSessionItemInterface {
   lastUpdatedTimestamp: string;
 }
 
-// export interface GetSessionsInterface {
-//   sessions: GetSessionItemInterface[];
-//   meta?: { totalItems: number };
-// }
-
-// export interface GetSessionInterface {
-//   sessionId: string;
-//   customerIdentity: string;
-//   channel: string;
-//   platform: string;
-//   ipAddresses: string[];
-//   locations: string[];
-//   deviceId: string;
-//   isIncognito: boolean;
-//   activeXEnabled: boolean;
-//   riskScore: string;
-//   lastEvent: {
-//     eventCode: string;
-//     eventName: string;
-//     transactionId: string;
-//     transactionAmount: string;
-//     maskedCard: string;
-//     phoneNumber: string;
-//     timestamp: string;
-//   };
-//   transactionSummary: {
-//     total: number;
-//     currency: string;
-//   };
-//   userAgent: string;
-//   timezone: string[];
-//   updatedAt: string;
-// }
 
 export interface GetSessionsListPayload {
   page: number;
@@ -90,11 +58,6 @@ export interface EventItem {
   ip: string;
   country: string;
   city: string;
-}
-
-export interface PaginationMeta {
-  totalPages: number;
-  totalItems: number;
 }
 
 export interface EventsData {
